Removed Ubuntu section due to time constraints...
Ubuntu derivatives base their Live CD on a program called casper.
Casper looks for a partition with a particular label of a particular
format and will use it as a persistent partition if it can find it.  So,
much like how the `persistence` partition was created for Kali, it's
possible to create a `casper-rw` partition for Ubuntu.  Unfortunately,
once the Ubuntu ISO has been written to a thumb drive, that thumb
drive's partition table is no longer formatted properly and using
standard partitioning tools to add an additional partition appears to be
impossible.

So, the procedure for creating an Ubuntu Live USB Drive with persistent
storage requires partitioning the drive, copying files to key locations
in order to satisfy SYSLINUX, and then using SYSLINUX to make the drive
bootable.